fix(terminal): source LAYOUTS from layouts module not SplitView barrel (cycle 8)
LayoutSwitcher and usePaneShortcuts both imported LAYOUTS via the
SplitView component barrel. The data constant lives in
SplitView/layouts.ts; importing through a sibling component creates
an unintended coupling that breaks silently if the barrel ever
narrows its re-exports. Switch both to direct imports from the
canonical source.

Resolves F1 (LOW) from Claude Code Review comment #4433402157.

Codex-Verify-Status: pass
